Require correct binary builds to release
A package must now build for all architechtures in its
`debian/control` file to be released. If not, a yellow warning
is printed in the console, and the sync is skipped.

This is done by checking reading the achitechtures that that the
package builds for, the checking if that package shows up in each
binary `Pacakges` file. A small quirk is the filename is checked
for, beacuse the build script only gives us access to the git
reponame, not the build package name. But there should never be a
conflict there anyway.

Out build server only properly builds on amd65 and arm64, so we
filter to only test those architechtures at most.