Fix build definitions
each architechture that is checked for build now has a map to
for architechture name to build folder. Some of these are redundant
but necessary forr architechtures like `linux-any` or `any` where
the name of the architechture doesn't corespond with the folder
that the build binary package data is put into.

Any architechtures that arn't put into the map are ignored and not
guarenteed to build for a package to be released.