Require correct binary builds to release (#194)
* Require correct binary builds to release

A package must now build for all architechtures in its
`debian/control` file to be released. If not, a yellow warning
is printed in the console, and the sync is skipped.

This is done by checking reading the achitechtures that that the
package builds for, the checking if that package shows up in each
binary `Pacakges` file. A small quirk is the filename is checked
for, beacuse the build script only gives us access to the git
reponame, not the build package name. But there should never be a
conflict there anyway.

Out build server only properly builds on amd65 and arm64, so we
filter to only test those architechtures at most.

* Add warning if packaging is missing a .dsc file

These packages were already skipped anyway. This just makes the
error more obvious and skips code that used to run even if there
was no logical output possible.

* Fix build definitions

each architechture that is checked for build now has a map to
for architechture name to build folder. Some of these are redundant
but necessary forr architechtures like `linux-any` or `any` where
the name of the architechture doesn't corespond with the folder
that the build binary package data is put into.

Any architechtures that arn't put into the map are ignored and not
guarenteed to build for a package to be released.

* Add `--ignore-binary` flag

Add a flag to allow syncing packages that don't have a binary
built. This is useful if:
1. The binary check may have a bug in the future
2. We might want to release a source only package in the future?

This will still get the warning if the binary failed to build
before prompting to release a package